requirements.txt
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
Flask==3.1.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
playwright==1.55.0
pycparser==2.23
pyee==13.0.0
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.0
python-dotenv==1.1.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
